AtCoder.Convolution

Description

It calculates \((+,\times)\) convolution. Given two arrays \(a_0, a_1, \cdots, a_{N - 1}\) and \(b_0, b_1, \cdots, b_{M - 1}\), it calculates the array \(c\) of length \(N + M - 1\), defined by

\[ c_i = \sum_{j = 0}^i a_j b_{i - j} \]

Documentation

convolution :: forall p. (HasCallStack, Modulus p) => Vector (StaticModInt p) -> Vector (StaticModInt p) -> Vector (StaticModInt p)

Calculates the convolution in \(\bmod m\) for a vector of StaticModInt. It returns an empty array if at least one of \(a\) and \(b\) are empty.

Constraints

\(2 \leq m \leq 2 \times 10^9\)

\(m\) is prime.

There is an integer \(c\) with \(2^c | (m - 1)\) and \(|a| + |b| - 1 \leq 2^c\).

Complexity

\(O(n\log{n} + \log{\mathrm{mod}})\), where \(n = |a| + |b|\).

convolutionMod :: forall p a. (HasCallStack, Modulus p, Integral a, Unbox a) => Proxy p -> Vector a -> Vector a -> Vector a

convolution for any Integral a.

Constraints

\(2 \leq m \leq 2 \times 10^9\)

\(m\) is prime.

There is an integer \(c\) with \(2^c | (m - 1)\) and \(|a| + |b| - 1 \leq 2^c\).

Complexity

\(O(n\log{n} + \log{\mathrm{mod}})\), where \(n = |a| + |b|\).

convolution64 :: HasCallStack => Vector Int -> Vector Int -> Vector Int

Calculates the convolution. It returns an empty array if at least one of \(a\) and \(b\) are empty.

Constraints

\(|a| + |b| - 1 \leq 2^{24}\)

All the elements of the array are in Int after convolution

Complexity

\(O(n\log{n})\), where \(n = |a| + |b|\).