AtCoder.Extra.Math

Description

Extra math module.

Documentation

power :: Int -> (a -> a -> a) -> a -> a

\(O(\log n)\) Calculates \(s^n\) with custom multiplication operator using the binary lifting technique.

The internal implementation is taken from stimes, but powers uses strict evaluation and is often much faster.

Constraints

\(n \gt 0\)

mtimes' :: Monoid a => Int -> a -> a

\(O(\log n)\) Strict mtimes.

Constraints

\(n \ge 0\)