AtCoder.Extra.Monoid

Description

Extra module of pre-defined monoids and operators. SegAct and monoids in submodules are re-exported.

Documentation

class Monoid f => SegAct f a where

Haskell reprentation of the LazySegTree properties. User can implement either segAct or segActWithLength.

Constraints

Identity: \(\mathrm{segAct} \mathrm{mempty} a = a\).

Left monoid action: \((f_2 \diamond f_1) a = f_2 (f_1 a)\)

Endomorphism: \(f (a_1 \diamond a_2) = (f a_1) \diamond (f a_2)\)

Minimal complete definition

Nothing

Methods

segAct :: f -> a -> a

Lazy segment tree action.

segActWithLength :: Int -> f -> a -> a

Lazy segment tree action with target monoid length.

If you implement SegAt with this function, you don't have to store the monoid length, since it's given externally.

newtype Affine1 a

SegAct instance of one-dimensional affine transformation \(f: x \rightarrow a x + b\).

Composition and dual

\((f_1 \diamond f_2) v := (f_1 . f_2) v\). If you need foldr \([f_l, .., f_r]\) on a segment tree, be sure to wrap Affine1 with Dual.

Constructors

Affine1 (Affine1Repr a)

type Affine1Repr a = (a, a)

Affine1 internal representation. Tuples are not the fastest representation, but it's easier to implement Unbox.

newtype RangeSet a

SegAct instance of range set action.

Constructors

RangeSet (RangeSetRepr a)

newtype RangeSetId a

SegAct instance of range set action over ideomponent monoid.

Constructors

RangeSetId (RangeSetIdRepr a)