AtCoder.Internal.GrowVec

Description

Growable vector with some runtime overhead.

Documentation

data GrowVec s a

Growable vector with some runtime overhead.

Constructors

GrowVec

Fields

posGV :: !(MVector s Int)

Stores [l, r) range in the vecGV.

vecGV :: !(MutVar s (MVector s a))

new :: (PrimMonad m, Unbox a) => Int -> m (GrowVec (PrimState m) a)

\(O(n)\) Creates GrowVec with initial capacity \(n\).

build :: (PrimMonad m, Unbox a) => Vector a -> m (GrowVec (PrimState m) a)

\(O(n)\) Creates GrowVec with initial values.

reserve :: (HasCallStack, PrimMonad m, Unbox a) => GrowVec (PrimState m) a -> Int -> m ()

\(O(n)\) Reserves the internal storage capacity.

>>> growVec <- new @_ @Int 0
>>> capacity growVec
0
>>> reserve growVec 20
>>> capacity growVec
20

read :: (HasCallStack, PrimMonad m, Unbox a) => GrowVec (PrimState m) a -> Int -> m a

\(O(1)\) Yields the element at the given position. Will throw an exception if the index is out of range.

write :: (HasCallStack, PrimMonad m, Unbox a) => GrowVec (PrimState m) a -> Int -> a -> m ()

\(O(1)\) Yields the element at the given position. Will throw an exception if the index is out of range.

pushBack :: (PrimMonad m, Unbox a) => GrowVec (PrimState m) a -> a -> m ()

Amortized \(O(1)\). Grow the capacity twice

popBack :: (PrimMonad m, Unbox a) => GrowVec (PrimState m) a -> m (Maybe a)

\(O(1)\) Removes the last element from the buffer and returns it, or Nothing if it is empty.

popBack_ :: (PrimMonad m, Unbox a) => GrowVec (PrimState m) a -> m ()

\(O(1)\) popBack with return value discarded.

length :: (PrimMonad m, Unbox a) => GrowVec (PrimState m) a -> m Int

\(O(1)\) Returns the number of elements in the vector.

capacity :: (PrimMonad m, Unbox a) => GrowVec (PrimState m) a -> m Int

\(O(1)\) Returns the capacity of the internal the vector.

null :: (PrimMonad m, Unbox a) => GrowVec (PrimState m) a -> m Bool

\(O(1)\) Returns True if the vector is empty.

freeze :: (PrimMonad m, Unbox a) => GrowVec (PrimState m) a -> m (Vector a)

\(O(n)\) Yields an immutable copy of the mutable vector.

unsafeFreeze :: (PrimMonad m, Unbox a) => GrowVec (PrimState m) a -> m (Vector a)

\(O(1)\) Unsafely converts a mutable vector to an immutable one without copying. The mutable vector may not be used after this operation.