AtCoder.Internal.String

Documentation

saNaive :: HasCallStack => Vector Int -> Vector Int

\(O(n^2)\) Internal implementation of suffix array creation.

saDoubling :: HasCallStack => Vector Int -> Vector Int

\(O(n \log n)\) Internal implementation of suffix array creation.

saIsImpl :: HasCallStack => Int -> Int -> Vector Int -> Int -> Vector Int

\(O(n)\) Internal implementation of suffix array creation.

saIs :: HasCallStack => Vector Int -> Int -> Vector Int

\(O(n)\) Internal implementation of suffix array creation.

SA-IS, linear-time suffix array construction Reference: G. Nong, S. Zhang, and W. H. Chan, Two Efficient Algorithms for Linear Time Suffix Array Construction

saIsManual :: HasCallStack => Int -> Int -> Vector Int -> Int -> Vector Int

\(O(n)\) Internal implementation of suffix array creation.

SA-IS, linear-time suffix array construction Reference: G. Nong, S. Zhang, and W. H. Chan, Two Efficient Algorithms for Linear Time Suffix Array Construction