AtCoder.LazySegTree

Description

Lazily propagted segment tree. It is the data structure for the pair of a monoid \((S, \cdot: S \times S \to S, e \in S)\) and a set \(F\) of \(S \to S\) mappings that satisfies the following properties.

\(F\) contains the identity map \(\mathrm{id}\), where the identity map is the map that satisfies \(\mathrm{id}(x) = x\) for all \(x \in S\).

\(F\) is closed under composition, i.e., \(f \circ g \in F\) holds for all \(f, g \in F\).

\(f(x \cdot y) = f(x) \cdot f(y)\) holds for all \(f \in F\) and \(x, y \in S\).

Given an array \(S\) of length \(N\), it processes the following queries in \(O(\log N)\) time.

Acting the map \(f\in F\) (cf. \(x = f(x)\)) on all the elements of an interval

Calculating the product of the elements of an interval

For simplicity, in this document, we assume that the oracles op, e, mapping, composition, and id work in constant time. If these oracles work in \(O(T)\) time, each time complexity appear in this document is multipled by \(O(T)\).

Storing boxed types

If you really need to store boxed types to LazySegTree, use DoNotUnboxStrict a or other wrappers.

Example

Usage

Here, we're using Affine1 as monoid action and Sum as acted monoid.

>>> import AtCoder.LazySegTree qualified as LST
>>> import AtCoder.Extra.Monoid (SegAct(..), Affine1(..))
>>> import Data.Semigroup (Sum(..))
>>> seg <- LST.build @_ @(Affine1 Int) @(Sum Int) $ VU.fromList [1, 2, 3, 4]
>>> LST.applyIn seg 1 3 $ Affine1 (2, 1) -- \x -> 2 * x + 1
>>> LST.write seg 3 $ Sum 10 -- [1, 5, 7, 10]
>>> LST.modify seg (+ 1) 0   -- [2, 5, 7, 10]
>>> LST.read seg 1
Sum {getSum = 5}
>>> LST.prod seg 0 3
Sum {getSum = 14}
>>> LST.allProd seg
Sum {getSum = 24}
>>> LST.maxRight seg 0 (<= (Sum 10)) -- sum [0, 2) = 7 <= 10
2
>>> LST.minLeft seg 4 (<= (Sum 10)) -- sum [3, 4) = 10 <= 10
3

SegAct instance

LazeSegTree functions require SegAct instance. Take Affine1 as an example.

{-# LANGUAGE TypeFamilies #-}

import AtCoder.LazySegTree (SegAct (..)) -- also re-exported from AtCoder.Extra.Monoid.
import Data.Monoid
import Data.Vector.Generic qualified as VG
import Data.Vector.Generic.Mutable qualified as VGM
import Data.Vector.Unboxed qualified as VU
import Data.Vector.Unboxed.Mutable qualified as VUM

newtype Affine1 a = Affine1 (Affine1 a)
  deriving newtype (Eq, Ord, Show)

type Affine1 a = (a, a)

instance (Num a) => Semigroup (Affine1 a) where
  {-# INLINE (<>) #-}
  (Affine1 (!a1, !b1)) <> (Affine1 (!a2, !b2)) = Affine1 (a1 * a2, a1 * b2 + b1)

instance (Num a) => Monoid (Affine1 a) where
  {-# INLINE mempty #-}
  mempty = Affine1 (1, 0)

instance (Num a) => SegAct (Affine1 a) (Sum a) where
  {-# INLINE segActWithLength #-}
  segActWithLength len (Affine1 (!a, !b)) !x = a * x + b * fromIntegral len

-- Derive Unbox:
newtype instance VU.MVector s (Affine1 a) = MV_Affine1 (VU.MVector s (Affine1 a))
newtype instance VU.Vector (Affine1 a) = V_Affine1 (VU.Vector (Affine1 a))
deriving instance (VU.Unbox a) => VGM.MVector VUM.MVector (Affine1 a)
deriving instance (VU.Unbox a) => VG.Vector VU.Vector (Affine1 a)
instance (VU.Unbox a) => VU.Unbox (Affine1 a)

Tips:

New monoid always come from the left: new <> old.

prod returns \(a_l \diamond f_{l + 1} \diamond .. \diamond a_{r - 1}\).

If you need \(a_{r - 1} \diamond a_{r - 2} \diamond .. \diamond f_{l}\), wrap your monoid in Dual.

Major changes from the original ac-library

The implementaion is Monoid and SegAct based.

get and set are renamed to read and write.

modify and modifyM are added.

Documentation

class Monoid f => SegAct f a where

Haskell reprentation of the LazySegTree properties. User can implement either segAct or segActWithLength.

Constraints

Identity: \(\mathrm{segAct} \mathrm{mempty} a = a\).

Left monoid action: \((f_2 \diamond f_1) a = f_2 (f_1 a)\)

Endomorphism: \(f (a_1 \diamond a_2) = (f a_1) \diamond (f a_2)\)

Minimal complete definition

Nothing

Methods

segAct :: f -> a -> a

Lazy segment tree action.

segActWithLength :: Int -> f -> a -> a

Lazy segment tree action with target monoid length.

If you implement SegAt with this function, you don't have to store the monoid length, since it's given externally.

Constraints

Linear monoid action: segActWithLength len f a = stimes len (segAct f a)} a.

data LazySegTree s f a

Lazy segment tree defined around SegAct.

Constructors

LazySegTree

Fields

nLst :: !Int

Valid length.

sizeLst :: !Int

\(\lceil \log_2 \mathrm{nLst} \rceil\)

logLst :: !Int

\(\log_2 \mathrm{sizeLst}\).

dLst :: !(MVector s a)

Data storage of length 2 * sizeLst.

lzLst :: !(MVector s f)

Data storage of length sizeLst.

new :: (HasCallStack, PrimMonad m, Monoid f, Unbox f, Monoid a, Unbox a) => Int -> m (LazySegTree (PrimState m) f a)

Creates an array of length n. All the elements are initialized to mempty.

Constraints

\(0 \leq n\)

Complexity

\(O(n)\)

build :: (PrimMonad m, Monoid f, Unbox f, Monoid a, Unbox a) => Vector a -> m (LazySegTree (PrimState m) f a)

Creates an array with initial values vs.

Constraints

\(0 \leq n\)

Complexity

\(O(n)\)

write :: (HasCallStack, PrimMonad m, SegAct f a, Unbox f, Monoid a, Unbox a) => LazySegTree (PrimState m) f a -> Int -> a -> m ()

Sets \(p\)-th value of the array to \(x\).

Constraints

\(0 \leq p \lt n\)

Complexity

\(O(\log n)\)

modify :: (HasCallStack, PrimMonad m, SegAct f a, Unbox f, Monoid a, Unbox a) => LazySegTree (PrimState m) f a -> (a -> a) -> Int -> m ()

(Extra API) Modifies \(p\)-th value of the array to \(x\).

Constraints

\(0 \leq p \lt n\)

Complexity

\(O(\log n)\)

modifyM :: (HasCallStack, PrimMonad m, SegAct f a, Unbox f, Monoid a, Unbox a) => LazySegTree (PrimState m) f a -> (a -> m a) -> Int -> m ()

(Extra API) Modifies \(p\)-th value of the array to \(x\).

Constraints

\(0 \leq p \lt n\)

Complexity

\(O(\log n)\)

read :: (HasCallStack, PrimMonad m, SegAct f a, Unbox f, Monoid a, Unbox a) => LazySegTree (PrimState m) f a -> Int -> m a

Returns \(p\)-th value of the array.

Constraints

\(0 \leq p \lt n\)

Complexity

\(O(\log n)\)

prod :: (HasCallStack, PrimMonad m, SegAct f a, Unbox f, Monoid a, Unbox a) => LazySegTree (PrimState m) f a -> Int -> Int -> m a

Returns the product of \([a[l], ..., a[r - 1]]\), assuming the properties of the monoid. It returns mempty if \(l = r\).

Constraints

\(0 \leq l \leq r \leq n\)

Complexity

\(O(\log n)\)

allProd :: (PrimMonad m, Monoid a, Unbox a) => LazySegTree (PrimState m) f a -> m a

Returns the product of \([op(a[0], ..., a[n - 1])]\), assuming the properties of the monoid. It returns mempty if \(n = 0\).

Complexity

\(O(1)\)

applyAt :: (HasCallStack, PrimMonad m, SegAct f a, Unbox f, Monoid a, Unbox a) => LazySegTree (PrimState m) f a -> Int -> f -> m ()

Applies segAct f to an index p.

Constraints

\(0 \leq p \lt n\)

Complexity

\(O(\log n)\)

applyIn :: (HasCallStack, PrimMonad m, SegAct f a, Unbox f, Monoid a, Unbox a) => LazySegTree (PrimState m) f a -> Int -> Int -> f -> m ()

Applies segAct f to an interval [l, r).

Constraints

\(0 \leq l \leq r \leq n\)

Complexity

\(O(\log n)\)

maxRight :: (HasCallStack, PrimMonad m, SegAct f a, Unbox f, Monoid a, Unbox a) => LazySegTree (PrimState m) f a -> Int -> (a -> Bool) -> m Int

Applies a binary search on the segment tree. It returns an index \(r\) that satisfies both of the followings.

\(r = l\) or \(g(a[l] \diamond a[l + 1] \diamond ... \diamond a[r - 1]))\) returns True.

\(r = n\) or \(g(a[l] \diamond a[l + 1] \diamond ... \diamond a[r]))\) returns False.

If \(g\) is monotone, this is the maximum \(r\) that satisfies \(g(a[l] \diamond a[l + 1] \diamond ... \diamond a[r - 1])\).

Constraints

If \(g\) is called with the same argument, it returns the same value, i.e., \(g\) has no side effect.

g mempty == True.

\(0 \leq l \leq n\)

Complexity

\(O(\log n)\)

maxRightM :: (HasCallStack, PrimMonad m, SegAct f a, Unbox f, Monoid a, Unbox a) => LazySegTree (PrimState m) f a -> Int -> (a -> m Bool) -> m Int

Monadic version of maxRight.

Constraints

If \(g\) is called with the same argument, it returns the same value, i.e., \(g\) has no side effect.

g mempty == True.

\(0 \leq l \leq n\)

Complexity

\(O(\log n)\)

minLeft :: (HasCallStack, PrimMonad m, SegAct f a, Unbox f, Monoid a, Unbox a) => LazySegTree (PrimState m) f a -> Int -> (a -> Bool) -> m Int

Applies a binary search on the segment tree. It returns an index \(l\) that satisfies both of the following.

\(l = r\) or \(g(a[l] \diamond a[l + 1] \diamond ... \diamond a[r - 1])\) returns True.

\(l = 0\) or \(g(a[l - 1] \diamond a[l] \diamond ... \diamond a[r - 1])\) returns False.

If \(g\) is monotone, this is the minimum \(l\) that satisfies \(g(a[l] \diamond a[l + 1] \diamond ... \diamond a[r - 1])\).

Constraints

if \(g\) is called with the same argument, it returns the same value, i.e., \(g\) has no side effect.

g mempty == True.

\(0 \leq r \leq n\)

Complexity

\(O(\log n)\)

minLeftM :: (HasCallStack, PrimMonad m, SegAct f a, Unbox f, Monoid a, Unbox a) => LazySegTree (PrimState m) f a -> Int -> (a -> m Bool) -> m Int

Monadic version of minLeft.

Constraints

if \(g\) is called with the same argument, it returns the same value, i.e., \(g\) has no side effect.

g mempty == True.

\(0 \leq r \leq n\)

Complexity

\(O(\log n)\)