AtCoder.MaxFlow

Description

It solves maximum flow problem.

Example

>>> import AtCoder.MaxFlow qualified as MF
>>> g <- MF.new @_ @Int 3        -- 0   1   2
>>> MF.addEdge g 0 1 (2 :: Int)  -- 0-->1   2
0
>>> MF.addEdge_ g 1 2 (1 :: Int) -- 0-->1-->2
>>> MF.flow g 0 2 maxBound -- MF.maxFlow g 0 2
1
>>> MF.getEdge g 0 -- returns (from, to, cap, flow)
(0,1,2,1)
>>> MF.minCut g 0 -- returns a bit vector
[1,1,0]

Documentation

data MfGraph s cap

Max flow graph.

new :: (PrimMonad m, Unbox cap) => Int -> m (MfGraph (PrimState m) cap)

Creates a graph of n vertices and \(0\) edges. cap is the type of the capacity.

Constraints

\(0 \leq n\)

Complexity

\(O(n)\)

addEdge :: (HasCallStack, PrimMonad m, Num cap, Ord cap, Unbox cap) => MfGraph (PrimState m) cap -> Int -> Int -> cap -> m Int

Adds an edge oriented from the vertex from to the vertex to with the capacity cap and the flow amount \(0\). It returns an integer \(k\) such that this is the \(k\)-th edge that is added.

Constraints

\(0 \leq \mathrm{from}, \mathrm{to} \lt n\)

\(0 \leq \mathrm{cap}\)

Complexity

\(O(1)\) amortized

addEdge_ :: (HasCallStack, PrimMonad m, Num cap, Ord cap, Unbox cap) => MfGraph (PrimState m) cap -> Int -> Int -> cap -> m ()

addEdge with return value discarded.

Constraints

\(0 \leq \mathrm{from}, \mathrm{to} \lt n\)

\(0 \leq \mathrm{cap}\)

Complexity

\(O(1)\) amortized

getEdge :: (HasCallStack, PrimMonad m, Num cap, Ord cap, Unbox cap) => MfGraph (PrimState m) cap -> Int -> m (Int, Int, cap, cap)

\(O(1)\) Returns the current internal state of \(i\)-th edge: (from, to, cap, flow). The edges are ordered in the same order as added by addEdge.

Constraints

\(0 \leq i \lt m\)

Complexity

\(O(1)\)

edges :: (PrimMonad m, Num cap, Ord cap, Unbox cap) => MfGraph (PrimState m) cap -> m (Vector (Int, Int, cap, cap))

Returns the current internal state of the edges: (from, to, cap, flow). The edges are ordered in the same order as added by addEdge.

Complexity

\(O(m)\), where \(m\) is the number of added edges.

changeEdge :: (HasCallStack, PrimMonad m, Num cap, Ord cap, Unbox cap) => MfGraph (PrimState m) cap -> Int -> cap -> cap -> m ()

\(O(1)\) Changes the capacity and the flow amount of the $i$-th edge to newCap and newFlow, respectively. It doesn't change the capacity or the flow amount of other edges.

Constraints

\(0 \leq \mathrm{newflow} \leq \mathrm{newcap}\)

Complexity

\(O(1)\)

flow :: (HasCallStack, PrimMonad m, Num cap, Ord cap, Unbox cap) => MfGraph (PrimState m) cap -> Int -> Int -> cap -> m cap

Augments the flow from \(s\) to \(t\) as much as possible, until reaching the amount of flowLimit. It returns the amount of the flow augmented. You may call it multiple times.

Constraints

\(s \neq t\)

\(0 \leq s, t \lt n\)

Complexity

\(O((n + m) \sqrt{m})\) (if all the capacities are \(1\)),

\(O(n^2 m)\) (general), or

\(O(F(n + m))\), where \(F\) is the returned value

maxFlow :: (HasCallStack, PrimMonad m, Num cap, Ord cap, Bounded cap, Unbox cap) => MfGraph (PrimState m) cap -> Int -> Int -> m cap

flow with no capacity limit.

Constraints

\(s \neq t\)

\(0 \leq s, t \lt n\)

Complexity

\(O((n + m) \sqrt{m})\) (if all the capacities are \(1\)),

\(O(n^2 m)\) (general), or

\(O(F(n + m))\), where \(F\) is the returned value

minCut :: (PrimMonad m, Num cap, Ord cap, Unbox cap) => MfGraph (PrimState m) cap -> Int -> m (Vector Bit)

Returns a vector of length \(n\), such that the \(i\)-th element is True if and only if there is a directed path from \(s\) to \(i\) in the residual network. The returned vector corresponds to a \(s-t\) minimum cut after calling flow(s, t) exactly once without flow_limit.

Complexity

\(O(n + m)\), where \(m\) is the number of added edges.