AtCoder.ModInt

Description

It is the struct that treats the modular arithmetic. All the remaining parts of AC Library works without modint, so you don't necessarily read this to use the remaining parts.

For most of the problems, it is sufficient to use ModInt998244353, ModInt1000000007, or modInt, which can be used as follows.

FIXME: write example

Major changes from the original ac-library

DynamicModInt is removed.

Documentation

class KnownNat a => Modulus a where

KnownNat with meta information used for modulus.

Minimal complete definition

isPrimeModulus

Methods

isPrimeModulus :: Proxy# a -> Bool

primitiveRootModulus :: Proxy# a -> Int

Be sure to override for the speed.

type ModInt998244353 = StaticModInt 998244353

type ModInt1000000007 = StaticModInt 1000000007

newtype StaticModInt a

Constructors

StaticModInt

Fields

unStaticModInt :: Word32

modVal :: forall a. KnownNat a => Proxy a -> Int

Retrieves Int from KnownNat.

>>> :set -XDataKinds
>>> import Data.Proxy (Proxy(..))
>>> import GHC.TypeLits (natVal)
>>> fromInteger (natVal (Proxy @42))
42

modVal# :: forall a. KnownNat a => Proxy# a -> Int

Retrieves Int from KnownNat.

>>> :set -XDataKinds
>>> :set -XMagicHash
>>> import GHC.Exts (proxy#)
>>> import GHC.TypeLits (natVal')
>>> fromInteger (natVal' (proxy# @42))
42

new :: forall a. KnownNat a => Int -> StaticModInt a

new32 :: forall a. KnownNat a => Word32 -> StaticModInt a

modulus :: forall a. KnownNat a => StaticModInt a -> Int

raw :: KnownNat a => Int -> StaticModInt a

raw32 :: KnownNat a => Word32 -> StaticModInt a

val :: KnownNat a => StaticModInt a -> Int

val32 :: KnownNat a => StaticModInt a -> Word32

pow :: (HasCallStack, KnownNat a) => StaticModInt a -> Int -> StaticModInt a

inv :: forall a. (HasCallStack, Modulus a) => StaticModInt a -> StaticModInt a