AtCoder.Convolution

Description

It calculates \((+,\times)\) convolution. Given two arrays \(a_0, a_1, \cdots, a_{N - 1}\) and \(b_0, b_1, \cdots, b_{M - 1}\), it calculates the array \(c\) of length \(N + M - 1\), defined by

\[ c_i = \sum_{j = 0}^i a_j b_{i - j} \]

Example

>>> import AtCoder.Convolution qualified as C
>>> import AtCoder.ModInt qualified as M
>>> import Data.Proxy (Proxy)
>>> import Data.Vector.Unboxed qualified as VU
>>> type Mint = M.ModInt998244353
>>> let modInt :: Int -> Mint; modInt = M.new
>>> 
>>> -- Convolution for mod int:
>>> let a = VU.map modInt $ VU.fromList [1, 2, 3, 4]
>>> let b = VU.map modInt $ VU.fromList [5, 6, 7, 8]
>>> C.convolution a b
[5,16,34,60,61,52,32]
>>> 
>>> -- Convolution for any `Integral a`:
>>> let a = VU.fromList @Int [1, 2, 3, 4]
>>> let b = VU.fromList @Int [5, 6, 7, 8]
>>> C.convolutionMod (Proxy @998244353) a b
[5,16,34,60,61,52,32]

Documentation

convolution :: forall (p :: Nat). (HasCallStack, Modulus p) => Vector (ModInt p) -> Vector (ModInt p) -> Vector (ModInt p)

Calculates the convolution in \(\bmod m\) for a vector of ModInt. It returns an empty array if at least one of \(a\) and \(b\) are empty.

Constraints

\(2 \leq m \leq 2 \times 10^9\)

\(m\) is prime.

There is an integer \(c\) with \(2^c | (m - 1)\) and \(|a| + |b| - 1 \leq 2^c\).

Complexity

\(O(n\log{n} + \log{\mathrm{mod}})\), where \(n = |a| + |b|\).

convolutionMod :: forall (p :: Nat) a. (HasCallStack, Modulus p, Integral a, Unbox a) => Proxy p -> Vector a -> Vector a -> Vector a

convolution for any Integral a.

Constraints

\(2 \leq m \leq 2 \times 10^9\)

\(m\) is prime.

There is an integer \(c\) with \(2^c | (m - 1)\) and \(|a| + |b| - 1 \leq 2^c\).

Complexity

\(O(n\log{n} + \log{\mathrm{mod}})\), where \(n = |a| + |b|\).

convolution64 :: HasCallStack => Vector Int -> Vector Int -> Vector Int

Calculates the convolution. It returns an empty array if at least one of \(a\) and \(b\) are empty.

Constraints

\(|a| + |b| - 1 \leq 2^{24}\)

Complexity

\(O(n\log{n})\), where \(n = |a| + |b|\).