AtCoder.Extra.Monoid

Description

Extra module of pre-defined monoids and operators.

Documentation

class Monoid f => SegAct f a where

Haskell reprentation of the above properties.

Minimal complete definition

segAct

Methods

segAct :: f -> a -> a

Lazy segment tree action.

Constraints

Left monoid action: \((f_2 \diamond f_1) a = f_2 (f_1 a)\)

Endomorphism: \(f (a_1 \diamond a_2) = (f a_1) \diamond (f a_2)\)

segActWithLength :: Int -> f -> a -> a

Lazy segment tree action with target monoid length.

By default implementation, it discards the monoid length and falls back to segAct. If you implement this function, you don't have to store the monoid length, since it's given externally. segAct can be implemented with segActWithLength 1.

newtype Affine2d a

2D affine transformation \(f: x \rightarrow a x + b\).

The acted target type is V2, which holds the length at the second element.

Composition and dual

\((f_1 \diamond f_2) v := (f_1 . f_2) v\). If yo need foldr [f_l, .., f_r] on segment tree, be sure to wrap Affine2d with Dual.

Constructors

Affine2d (Affine2dRepr a)