AtCoder.Internal.MinHeap

Description

Minimum binary heap. Mutable and fixed-sized.

https://en.wikipedia.org/wiki/Binary_heap

Example

>>> heap <- new @Int 4
>>> capacity heap
4
>>> push heap 10
>>> push heap 0
>>> push heap 5
>>> length heap
3
>>> pop heap
Just 0
>>> peek heap
Just 5
>>> pop heap
Just 5
>>> clear heap
>>> null heap
True

Documentation

data Heap s a

Minimum binary heap. Mutable and fixed-sized.

Indices are zero-based.

    0
  1   2
 3 4 5 6

INVARIANT (min heap): child values are bigger than or equal to their parent value.

Constructors

Heap

Fields

sizeBH_ :: !(MVector s Int)

Size of the heap.

dataBH :: !(MVector s a)

Storage.

new :: (Unbox a, PrimMonad m) => Int -> m (Heap (PrimState m) a)

\(O(n)\) Creates Heap with capacity \(n\).

capacity :: Unbox a => Heap s a -> Int

\(O(1)\) Returns the number of elements the heap can hold.

length :: (Unbox a, PrimMonad m) => Heap (PrimState m) a -> m Int

\(O(1)\) Returns the number of elements in the heap.

null :: (Unbox a, PrimMonad m) => Heap (PrimState m) a -> m Bool

\(O(1)\) Returns True if the heap is empty.

clear :: (Unbox a, PrimMonad m) => Heap (PrimState m) a -> m ()

\(O(1)\) Sets the length to zero.

push :: (HasCallStack, Ord a, Unbox a, PrimMonad m) => Heap (PrimState m) a -> a -> m ()

\(O(\log n)\) Inserts an element to the heap.

peek :: (Unbox a, PrimMonad m) => Heap (PrimState m) a -> m (Maybe a)

\(O(1)\) Returns the smallest value in the heap, or Nothing if it is empty.

pop :: (HasCallStack, Ord a, Unbox a, PrimMonad m) => Heap (PrimState m) a -> m (Maybe a)

\(O(\log n)\) Removes the last element from the heap and returns it, or Nothing if it is empty.

pop_ :: (HasCallStack, Ord a, Unbox a, PrimMonad m) => Heap (PrimState m) a -> m ()

\(O(\log n)\) Removes the last element from the heap and discards it.