AtCoder.Internal.Queue

Description

Queue with fixed size capacity.

Example

>>> que <- new @_ @Int 3
>>> capacity que
3
>>> pushBack que 0
>>> pushBack que 1
>>> pushBack que 2
>>> length que
3
>>> popFront que
Just 0
>>> freeze que
[1,2]
>>> pushFront que 10
>>> pushFront que 1000
*** Exception: AtCoder.Internal.Queue.pushFront: no empty front space
...
>>> unsafeFreeze que
[10,1,2]
>>> clear que
>>> pushBack que 0
>>> pushBack que 1
>>> pushBack que 2
>>> freeze que
[0,1,2]

Documentation

data Queue s a

Queue with fixed size capacity.

Constructors

Queue

Fields

posQ :: !(MVector s Int)

Stores [l, r) range in the vecQ.

vecQ :: !(MVector s a)

new :: (PrimMonad m, Unbox a) => Int -> m (Queue (PrimState m) a)

\(O(n)\) Creates Queue with capacity \(n\).

pushBack :: (HasCallStack, PrimMonad m, Unbox a) => Queue (PrimState m) a -> a -> m ()

\(O(1)\) Appends an element to the back. Will throw an exception if the index is out of range.

pushFront :: (HasCallStack, PrimMonad m, Unbox a) => Queue (PrimState m) a -> a -> m ()

\(O(1)\) Appends an element to the back. Will throw an exception if the index is out of range.

popFront :: (PrimMonad m, Unbox a) => Queue (PrimState m) a -> m (Maybe a)

\(O(1)\) Removes the first element from the queue and returns it, or Nothing if it is empty.

capacity :: Unbox a => Queue s a -> Int

\(O(1)\) Returns the array size.

length :: (PrimMonad m, Unbox a) => Queue (PrimState m) a -> m Int

\(O(1)\) Returns the number of elements in the queue.

null :: (PrimMonad m, Unbox a) => Queue (PrimState m) a -> m Bool

\(O(1)\) Returns True if the buffer is empty.

clear :: (PrimMonad m, Unbox a) => Queue (PrimState m) a -> m ()

\(O(1)\) Sets the length to zero.

freeze :: (PrimMonad m, Unbox a) => Queue (PrimState m) a -> m (Vector a)

\(O(n)\) Yields an immutable copy of the mutable vector.

unsafeFreeze :: (PrimMonad m, Unbox a) => Queue (PrimState m) a -> m (Vector a)

\(O(1)\) Unsafely converts a mutable vector to an immutable one without copying. The mutable vector may not be used after this operation.