AtCoder.Internal.String

Description

Internal implementation of String module.

Suffix array

saNaive :: HasCallStack => Vector Int -> Vector Int

\(O(n^2)\) Internal implementation of suffix array creation (naive).

Since: 1.0.0

saDoubling :: HasCallStack => Vector Int -> Vector Int

\(O(n \log n)\) Internal implementation of suffix array creation (doubling).

Since: 1.0.0

saIsImpl :: HasCallStack => Int -> Int -> Vector Int -> Int -> Vector Int

\(O(n)\) Internal implementation of suffix array creation (suffix array induced sorting).

Since: 1.0.0

saIs :: HasCallStack => Vector Int -> Int -> Vector Int

\(O(n)\) Internal implementation of suffix array creation (suffix array induced sorting).

SA-IS, linear-time suffix array construction. Reference: G. Nong, S. Zhang, and W. H. Chan, Two Efficient Algorithms for Linear Time Suffix Array Construction

Since: 1.0.0

saIsManual :: HasCallStack => Int -> Int -> Vector Int -> Int -> Vector Int

\(O(n)\) Internal implementation of suffix array creation (suffix array induced sorting).

SA-IS, linear-time suffix array construction. Reference: G. Nong, S. Zhang, and W. H. Chan, Two Efficient Algorithms for Linear Time Suffix Array Construction

Since: 1.0.0