AtCoder.ModInt

Description

It is the struct that treats the modular arithmetic. All the remaining parts of AC Library works without modint, so you don't necessarily read this to use the remaining parts.

For most of the problems, it is sufficient to use ModInt998244353, ModInt1000000007, which can be used as follows.

>>> import AtCoder.ModInt qualified as M
>>> type Mint = M.ModInt998244353
>>> let modInt :: Int -> Mint; modInt = M.new
>>> modInt 1000000000
1755647

Major changes from the original ac-library

DynamicModInt is removed.

Since: 1.0.0

Modulus

class KnownNat a => Modulus (a :: Nat) where

KnownNat with meta information used for modulus.

Since: 1.0.0

Minimal complete definition

isPrimeModulus

Methods

isPrimeModulus :: Proxy# a -> Bool

Returns if the modulus is a prime value.

Since: 1.0.0

primitiveRootModulus :: Proxy# a -> Int

Returns the primitive root of the modulus value. Note that the default implementation is slow.

Since: 1.0.0

type ModInt998244353 = ModInt 998244353

ModInt with modulus value 998244353.

Since: 1.0.0

type ModInt1000000007 = ModInt 1000000007

ModInt with modulus value 1000000007.

Since: 1.0.0

Helpers

modVal :: forall (a :: Nat). KnownNat a => Proxy a -> Int

Retrieves Int from KnownNat.

>>> import Data.Proxy (Proxy(..))
>>> modVal (Proxy @42)
42

Since: 1.0.0

modVal# :: forall (a :: Nat). KnownNat a => Proxy# a -> Int

Retrieves Int from KnownNat.

>>> :set -XMagicHash
>>> import GHC.Exts (proxy#)
>>> modVal# (proxy# @42)
42

Since: 1.0.0

ModInt

newtype ModInt (a :: k)

Word32 value that treats the modula arithmetic.

Constructors

ModInt

Fields

unModInt :: Word32

Constructors

Safe constructors

new :: forall (a :: Nat). KnownNat a => Int -> ModInt a

Creates ModInt from an Int value taking mod.

Since: 1.0.0

new32 :: forall (a :: Nat). KnownNat a => Word32 -> ModInt a

Creates ModInt from a Word32 value taking mod.

Since: 1.0.0

new64 :: forall (a :: Nat). KnownNat a => Word32 -> ModInt a

Creates ModInt from a Word64 value taking mod.

Since: 1.0.0

Unsafe constructor

unsafeNew :: forall (a :: Nat). KnownNat a => Word32 -> ModInt a

Creates ModInt without taking mod. It is the function for constant-factor speedup.

Constraints

\(0 \leq x \lt \mathrm{mod}\) (not asserted at runtime)

Since: 1.0.0

Accessors

Modulus value

modulus :: forall (a :: Nat). KnownNat a => ModInt a -> Int

Retrieve the mod from a ModInt object.

Complecity

\(O(1)\)

Since: 1.0.0

Internal value

val :: forall (a :: Nat). KnownNat a => ModInt a -> Int

Returns the internal value converted to Int.

Complecity

\(O(1)\)

Since: 1.0.0

val32 :: forall (a :: Nat). KnownNat a => ModInt a -> Word32

Returns the internal value as Word32 without type conversion. It is the function for constant-factor speedup.

Complecity

\(O(1)\)

Since: 1.0.0

val64 :: forall (a :: Nat). KnownNat a => ModInt a -> Word32

Returns the internal value converted to Word32.

Complecity

\(O(1)\)

Since: 1.0.0

Operators

pow :: forall (a :: Nat). (HasCallStack, KnownNat a) => ModInt a -> Int -> ModInt a

Returns \(x^n\). The implementation is a bit more efficient than ^.

Constraints

\(0 \le n\)

Complexity

\(O(\log n)\)

Since: 1.0.0

inv :: forall (a :: Nat). (HasCallStack, Modulus a) => ModInt a -> ModInt a

Returns \(y\) with \(xy \equiv 1\).

Constraints

gcd(val x, modulus x) == 1.

Complexity

\(O(\log \mathrm{mod})\)

Since: 1.0.0